WIP: Updating the driver
