+++ .github/CODEOWNERS
@@ -1,0 +1,1 @@
+*       @ssbarnea
