--- .github/labels.yml
@@ -1,38 +1,0 @@
-# Format and labels used aim to match those used by Ansible project
-# https://github.com/marketplace/actions/github-labeler
-- name: bug
-  color: "fbca04"
-  description: "This issue/PR relates to a bug."
-- name: deprecated
-  color: "fef2c0"
-  description: "This issue/PR relates to a deprecated module."
-- name: docs
-  color: "4071a5"
-  description: "This issue/PR relates to or includes documentation."
-- name: enhancement
-  color: "ededed"
-  description: "This issue/PR relates to a feature request."
-- name: feature
-  color: "006b75"
-  description: "This issue/PR relates to a feature request."
-- name: major
-  color: "c6476b"
-  description: "Marks an important and likely breaking change."
-- name: packaging
-  color: "4071a5"
-  description: "Packaging category"
-- name: performance
-  color: "555555"
-  description: "Relates to product or testing performance."
-- name: skip-changelog
-  color: "eeeeee"
-  description: "Can be missed from the changelog."
-- name: stale
-  color: "eeeeee"
-  description: "Not updated in long time, will be closed soon."
-- name: wontfix
-  color: "eeeeee"
-  description: "This will not be worked on"
-- name: test
-  color: "0e8a16"
-  description: "This PR relates to tests, QA, CI."
