+++ .github/workflows/push.yml
@@ -1,0 +1,12 @@
+# See https://github.com/ansible-community/devtools/blob/main/.github/workflows/push.yml
+name: push
+on:
+  push:
+    branches:
+      - main
+      - 'releases/**'
+      - 'stable/**'
+
+jobs:
+  push:
+    uses: ansible-community/devtools/.github/workflows/push.yml@main
