+++ .github/workflows/tox.yml
@@ -1,0 +1,70 @@
+name: tox
+
+on:
+  create:  # is used for publishing to TestPyPI
+    tags:  # any tag regardless of its name, no branches
+      - "**"
+  push:  # only publishes pushes to the main branch to TestPyPI
+    branches:  # any integration branch but not tag
+      - "main"
+  pull_request:
+  release:
+    types:
+      - published  # It seems that you can publish directly without creating
+  schedule:
+    - cron: 1 0 * * *  # Run daily at 0:01 UTC
+
+jobs:
+  build:
+    name: ${{ matrix.tox_env }}
+    # ubuntu-latest is current older and we need last podman toolkit.
+    # https://docs.github.com/en/actions/reference/virtual-environments-for-github-hosted-runners
+    runs-on: ubuntu-20.04
+    strategy:
+      fail-fast: false
+      matrix:
+        include:
+          - tox_env: lint
+          # - tox_env: docs
+          - tox_env: py36-ansible29
+            PREFIX: PYTEST_REQPASS=5
+          - tox_env: py36-ansible210
+            PREFIX: PYTEST_REQPASS=5
+          - tox_env: py36-ansible211
+            PREFIX: PYTEST_REQPASS=5
+          - tox_env: py37-ansible211
+            PREFIX: PYTEST_REQPASS=5
+          - tox_env: py38-ansible211
+            PREFIX: PYTEST_REQPASS=5
+          - tox_env: py39-ansible211
+            PREFIX: PYTEST_REQPASS=5
+          - tox_env: py39-devel-ansible211
+            PREFIX: PYTEST_REQPASS=5
+          - tox_env: packaging
+
+    steps:
+      - uses: actions/checkout@v1
+      - name: Find python version
+        id: py_ver
+        shell: python
+        if: ${{ contains(matrix.tox_env, 'py') }}
+        run: |
+          v = '${{ matrix.tox_env }}'.split('-')[0].lstrip('py')
+          print('::set-output name=version::{0}.{1}'.format(v[0],v[1:]))
+      # Even our lint and other envs need access to tox
+      - name: Install a default Python
+        uses: actions/setup-python@v2
+        if: ${{ ! contains(matrix.tox_env, 'py') }}
+      # Be sure to install the version of python needed by a specific test, if necessary
+      - name: Set up Python version
+        uses: actions/setup-python@v2
+        if: ${{ contains(matrix.tox_env, 'py') }}
+        with:
+          python-version: ${{ steps.py_ver.outputs.version }}
+      - name: Install dependencies
+        run: |
+          python -m pip install -U "pip>=21.2.4" "tox>=3.24.2"
+      - name: Run tox -e ${{ matrix.tox_env }}
+        run: |
+          echo "${{ matrix.PREFIX }} tox -e ${{ matrix.tox_env }}"
+          ${{ matrix.PREFIX }} tox -e ${{ matrix.tox_env }}
