--- MANIFEST.in
@@ -1,5 +1,0 @@
-include LICENSE
-include README.rst
-
-recursive-exclude * __pycache__
-recursive-exclude * *.py[co]
