--- molecule_digitalocean/cookiecutter/{{cookiecutter.molecule_directory}}/{{cookiecutter.scenario_name}}/prepare.yml
@@ -1,13 +1,0 @@
----
-{% raw -%}
-- name: Prepare
-  hosts: all
-  gather_facts: false
-  tasks:
-    - name: Install python for Ansible
-      raw: >-
-        test -e /usr/bin/python ||
-        (apt -y update && apt install -y python3 python3-pip)
-      become: true
-      changed_when: false
-{%- endraw %}
