+++ setup.cfg
@@ -1,16 +1,13 @@
 [aliases]
 dists = clean --all sdist bdist_wheel
 
-[bdist_wheel]
-universal = 1
-
 [metadata]
 name = molecule-digitalocean
 url = https://github.com/pycontribs/molecule-digitalocean
 project_urls =
     Bug Tracker = https://github.com/pycontribs/molecule-digitalocean/issues
     Release Management = https://github.com/pycontribs/molecule-digitalocean/projects
-    CI: Travis = https://travis-ci.com/pycontribs/molecule-digitalocean
+    CI = https://github.com/ansible-community/molecule-digitalocean/actions
     Source Code = https://github.com/pycontribs/molecule-digitalocean
 description = digitalocean Molecule Plugin :: run molecule tests on digitalocean
 long_description = file: README.rst
@@ -31,12 +28,12 @@ classifiers =
     License :: OSI Approved :: MIT License
     Natural Language :: English
     Operating System :: OS Independent
-    Programming Language :: Python :: 2
-    Programming Language :: Python :: 2.7
     Programming Language :: Python :: 3
     Programming Language :: Python :: 3.5
     Programming Language :: Python :: 3.6
     Programming Language :: Python :: 3.7
+    Programming Language :: Python :: 3.8
+    Programming Language :: Python :: 3.9
 
     Topic :: System :: Systems Administration
     Topic :: Utilities
@@ -51,7 +48,9 @@ keywords =
 
 [options]
 use_scm_version = True
-python_requires = >=2.7,!=3.0.*,!=3.1.*,!=3.2.*,!=3.3.*,!=3.4.*
+python_requires = >=3.6
+package_dir =
+  = src
 packages = find:
 include_package_data = True
 zip_safe = False
@@ -63,7 +62,7 @@ setup_requires =
 
 # These are required in actual runtime:
 install_requires =
-    molecule >= 3.0a0
+    molecule >= 3.5.0a0
     pyyaml >= 5.1, < 6
 
 [options.extras_require]
@@ -84,4 +83,4 @@ molecule.driver =
     digitalocean = molecule_digitalocean.driver:DigitalOcean
 
 [options.packages.find]
-where = .
+where = src
