+++ src/molecule_digitalocean/cookiecutter/cookiecutter.json
