+++ src/molecule_digitalocean/cookiecutter/{{cookiecutter.molecule_directory}}/{{cookiecutter.scenario_name}}/prepare.yml
@@ -1,0 +1,11 @@
+---
+- name: Prepare
+  hosts: all
+  gather_facts: false
+  tasks:
+    - name: Install python for Ansible
+      ansible.builtin.raw: >-
+        test -e /usr/bin/python3 ||
+        (apt -y update && apt install -y python3 python3-pip)
+      become: true
+      changed_when: false
