+++ src/molecule_digitalocean/test/scenarios/driver/digitalocean/molecule/default/playbook.yml
