+++ src/molecule_digitalocean/test/scenarios/driver/digitalocean/molecule/multi-node/molecule.yml
